Added ScrollView and changed rendering
The ScrollView is there so that I can (later) handle resized screens, as
well as handle screens that are too small.

The rendering doesn't need to have a bunch of objects that have to be
deleted and created, so I changed them to instance variables. I suppose
I could go with static methods, but then I lose the polymorphism.